Refactor app to separate logic in various modules
