+++ spotKeys/speech.py
@@ -1,0 +1,13 @@
+"""Stores speech-related functionality."""
+
+import tolk
+
+
+def initialize() -> None:
+	"""Initializes Tolk."""
+	tolk.load()
+
+
+def say(text: str) -> None:
+	"""Speaks the given text."""
+	tolk.speak(text)
